Wait for the configuration to be initialized
Map will be available before it is initialized resulting in unexpected behaviors. For example setRegionAnimated provides a slightly different result, but adding a 500ms delay it draws as expected (or adding this push also fixes it).

Wait for the map to be initialized before providing it to the context ensures the map acts as expected.